+++ RepositoriesAndData/PostRepository.cs
@@ -110,9 +110,9 @@ namespace RepositoriesAndData
             _connection.Open();
             SqliteCommand command = _connection.CreateCommand();
             command.CommandText = @"SELECT COUNT(*) FROM posts";
-            int result = (int)command.ExecuteScalar();
+            long result = (long)command.ExecuteScalar();
             _connection.Close();
-            return result;
+            return (int)result;
         }
 
         public int GetTotalPages()
